Add Binary writer, change type to Float
Float is written into the binary format, and it's pretty absurd to write
out text STLs with more precision than that.  It often makes sense to
use more precision in calculations, but the Triangle type provided here
isn't really intended for internal use.

+++ src/Graphics/Formats/STL/Printer.hs
@@ -27,4 +27,4 @@ vertex :: Vector -> Doc
 vertex v = text "vertex" <+> v3 v
 
 v3 :: Vector -> Doc
-v3 (x, y, z) = hsep [double x, double y, double z]
+v3 (x, y, z) = hsep [float x, float y, float z]
